2D Array Creation
This update comes with the addition of many features for the log table; swapping of log records, sorting log records by date, finding the next log number, adding a record and removing a record.

There have also been additions to the create table function allowing it to create the 2D array to view the log details. This is still being worked on.