Working on the outputting
Now has the ability to search for records on a given date and return the IDs of them. Currently the system is being tested to be able to get all records on a given date in order to use this information to be able to output them to the screen.