Reorganising code
The code is being reorganised because there is some redundancy in the main file so I am trying to reduce it since the main file is so big. It would be useful to be able to split the main file into smaller files however this will require more work on understanding js.