Loading Wips
Wips can now be loaded to the wip table and can be edited and saved when used. The routing problem has been fixed by changing the root of the router for the main routes, which should be named wip view routes. Finally, more bugs have been identified. There is a lot of validation to do at the end of this project.