Showing the stitch log
The stitch log is now shown on the screen for the wip viewing but it needs to be updated in order to automatically create the new start and finish entries. Additionally, the system needs to be updated in order to save the modifications made to the year view so that the log can be updated when the schedule view is changed not manually through the file.