Input box
The table now uses input boxes and is updated when the change is made to the input boxes. However, this system doesn't feel as stable as the old system so this is something to look out for. In addition, the code needs to be cleaned as it is currently not very clear and a bit messy. The next step is to add in the spreadsheet view of the stitch records.