Added wip overview route
The wip overview html file and route has been created in order to be able to create the wip overview section as well as the specific wip section. However the main functionality of this still needs to be implemented although I have had an idea of how to further simplify the system, this idea I will discuss in an issue named "Routing simplification" or something similar.