Small configuration update
The cors settings have been updated to fix one or two of the problems but there are still issues when trying to connect from another device to the server hosted on the computer. Additionally, if the server is hosted from the computer then the hostname parameter should be removed from the listen function in order to set the IP of the system to be the destination.