Month View
Started working on the month view of the wips. All of the wips in a given month are outputted. The year and month still needs to be able to be inputted by the user. The filling in of the table still needs to be done. Finally, there have been some changes to the routing of the site and these have been reflected to match the plan made in one of the issues.